correction if missing frames are in the borders

+++ extractBehaviorCameraTiming.py
@@ -69,7 +69,7 @@ for f in range(len(foldersRecordings)):
         # save data
         #idxToExclude = np.array([], dtype=np.int64)
         if assumePerfectRecording:
-            idxToExclude = np.array([], dtype=np.int64)
+            idxToExclude = np.array([])
             canBeUsed = True
         else:
             if (not erroneousFramesExist) and canBeUsed:
